requirements.txt
pandas==2.1.4
Pillow==10.1.0